Refactor iceberg multi-catalog smoke tests
